Allow for Replacer implementations
