+++ config/responsecache.php
@@ -43,4 +43,12 @@ return [
      * You may use a string or an array here.
      */
     'cache_tag' => '',
+
+    /*
+     * Here you may define replacers that dynamically replace content from the response.
+     * Each replacer must implement the ReplacerInterface.
+     */
+    'replacers' => [
+        \Spatie\ResponseCache\Replacers\CsrfTokenReplacer::class,
+    ],
 ];
