+++ src/Replacers/ReplacerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\ResponseCache\Replacers;
+
+interface ReplacerInterface
+{
+    public function getKey(): string;
+
+    public function getValue(): string;
+}
